+++ common/config.js
@@ -12,10 +12,18 @@ module.exports = {
   },{
     name: 'export-data'
   },{
+    name: 'gen-services'
+  },{
     name: 'export-swagger2-data'
   },{
     name: 'import-yapi-json'
   },{
     name: 'wiki'
-  }]
+  }, {
+    name: 'swagger-auto-sync'
+  }
+  // {
+  //   name: 'test'
+  // }
+]
 }