+++ server/controllers/base.js
@@ -39,7 +39,7 @@ class baseController {
 
     let openApiRouter = [
       '/api/open/run_auto_test',
-			'/api/open/import_data',
+      '/api/open/import_data',
 			'/api/interface/add',
 			'/api/interface/save',
 			'/api/interface/up',
@@ -47,13 +47,17 @@ class baseController {
 			'/api/interface/list',
 			'/api/interface/list_menu',
 			'/api/interface/add_cat',
-			'/api/interface/getCatMenu'
+      '/api/interface/getCatMenu',
+      '/api/interface/list_cat',
+      '/api/project/get',
+      '/api/plugin/export'
     ];
 
     let params = Object.assign({}, ctx.query, ctx.request.body);
     let token = params.token;
 
-    if (token && openApiRouter.indexOf(ctx.path) > -1) {
+    // 如果前缀是 /api/open，执行 parse token 逻辑
+    if (token && (openApiRouter.indexOf(ctx.path) > -1 || ctx.path.indexOf('/api/open/') === 0 )) {
       let tokens = parseToken(token)
 
       const oldTokenUid = '999999'
@@ -76,13 +80,14 @@ class baseController {
       //   }
       //   return (this.$tokenAuth = true);
       // }
-
+      
       let checkId = await this.getProjectIdByToken(token);
       if(!checkId){
         ctx.body = yapi.commons.resReturn(null, 42014, 'token 无效');
       }
       let projectData = await this.projectModel.get(checkId);
       if (projectData) {
+        ctx.query.pid = checkId; // 兼容：/api/plugin/export
         ctx.params.project_id = checkId;
         this.$tokenAuth = true;
         this.$uid = tokenUid;